backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
structlog==24.2.0

# Graph DB
neo4j==5.20.0

# Vector DB
qdrant-client==1.9.1

# Time-series (async HTTP)
httpx==0.27.0

# Terraform parsing
python-hcl2==4.3.2

# Embeddings / LLM
ollama==0.2.1

# MCP server
mcp==1.0.0

# Metrics exposition
prometheus-fastapi-instrumentator==7.0.0
